Allowing to lookup flows based on endpoint id.
Usually, if you just ask to connect to a remote ip:port over a certain transport, such as TCP, you don't
care which existing flow to use as long it is indeed connected to that remote host. So, you may have
multiple TCP connection connected to that remote host and therefore we should allow to use
any of those, which we now do.

Previously, an issue was that we created a new ConnectionId based on the local listening address
for a particular network interface but if it was TCP, then of course, there will never ever
be an outgoing connection from that listening port which meant we created a new flow
every time! That has been fixed.

Haven't fixed all unit tests yet but wanted to check it in so I don't loose it...